Update 2025-10-14-RUH037.md
so i was rereading the chapter on the main page and noticed there was an important line missing. I came to put it back in but it was in the text-- however there was a "SuandFriends0X" on the line directly above it? and other similar lines scattered through the text. 

It seems when there's an empty line between the code and the text it's safe, but if there's no empty line it somehow hides the text. might have to look into it